Refactor benchmark code for skqulacs
