+++ benchmarks/binary_classification_skqulacs.py
@@ -1,36 +1,37 @@
-import pandas as pd
+from typing import Tuple
+
 import numpy as np
+import pandas as pd
+from binary_classification_common import preprocess_input
+from numpy.random import Generator, default_rng
+from qulacs.gate import Pauli
 from sklearn import datasets
 from sklearn.metrics import f1_score
 from sklearn.model_selection import train_test_split
-from qulacs.gate import Pauli
-from numpy.random import default_rng, Generator
 
-from skqulacs.qnn import QNNClassifier
 from skqulacs.circuit import LearningCircuit
+from skqulacs.qnn import QNNClassifier
 
 
-def get_angles(x: np.ndarray) -> np.ndarray:
-    beta0 = 2 * np.arcsin(np.sqrt(x[1] ** 2) / np.sqrt(x[0] ** 2 + x[1] ** 2 + 1e-12))
-    beta1 = 2 * np.arcsin(np.sqrt(x[3] ** 2) / np.sqrt(x[2] ** 2 + x[3] ** 2 + 1e-12))
-    beta2 = 2 * np.arcsin(
-        np.sqrt(x[2] ** 2 + x[3] ** 2)
-        / np.sqrt(x[0] ** 2 + x[1] ** 2 + x[2] ** 2 + x[3] ** 2)
+def load_dataset() -> Tuple[np.ndarray, np.ndarray, np.ndarray, np.ndarray]:
+    iris = datasets.load_iris()
+    df = pd.DataFrame(iris.data, columns=iris.feature_names)
+    x = df.loc[:, ["petal length (cm)", "petal width (cm)"]]
+    x = x.to_numpy()
+    # Now we are going to solve binary classification, so exclude the third class.
+    index = iris.target != 2
+    x = preprocess_input(x[index])
+    y = iris.target[index]
+    x_train, x_test, y_train, y_test = train_test_split(
+        x, y, test_size=0.25, random_state=0
+    )
+    # Discard the half of the test data because in PennyLane implementation, 25% of data
+    # are used for validation, which is not needed for skqulacs's implementation.
+    _, x_test, _, y_test = train_test_split(
+        x_test, y_test, test_size=0.5, random_state=0
     )
 
-    return np.array([beta2, -beta1 / 2, beta1 / 2, -beta0 / 2, beta0 / 2])
-
-
-def preprocess_input(x: np.ndarray) -> np.ndarray:
-    padding = 0.3 * np.ones((len(x), 1))
-    X_pad = np.c_[np.c_[x, padding], np.zeros((len(x), 1))]
-
-    # normalize each input
-    normalization = np.sqrt(np.sum(X_pad ** 2, -1))
-    X_norm = (X_pad.T / normalization).T
-
-    # angles for state preparation are new features
-    return np.array([get_angles(x) for x in X_norm])
+    return x_train, x_test, y_train, y_test
 
 
 def create_circuit(depth: int) -> LearningCircuit:
@@ -76,22 +77,18 @@ def create_circuit(depth: int) -> LearningCircuit:
 
     return circuit
 
-def binary_classification_skqulacs(_n_qubit: int) -> float:
+
+def binary_classification_skqulacs(
+    circuit: LearningCircuit,
+    x_train: np.ndarray,
+    x_test: np.ndarray,
+    y_train: np.ndarray,
+    y_test: np.ndarray,
+) -> float:
     """Solve a binary classification problem for the subset of iris dataset.
 
     Returns: F1 score for trained model.
     """
-    iris = datasets.load_iris()
-    df = pd.DataFrame(iris.data, columns=iris.feature_names)
-    x = df.loc[:, ["petal length (cm)", "petal width (cm)"]]
-    x = x.to_numpy()
-    # Now we are going to solve binary classification, so exclude the third class.
-    index = iris.target != 2
-    x = preprocess_input(x[index])
-    y = iris.target[index]
-    x_train, x_test, y_train, y_test = train_test_split(
-        x, y, test_size=0.25, random_state=0
-    )
 
     num_class = 2
     circuit = create_circuit(6)
@@ -101,6 +98,6 @@ def binary_classification_skqulacs(_n_qubit: int) -> float:
     return f1_score(y_test, y_pred)
 
 
-if __name__ == "__main__":
-    score = binary_classification_skqulacs(2)
-    print(score)
+# if __name__ == "__main__":
+#     score = binary_classification_skqulacs()
+#     print(score)
